system_agent/requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.23.0
babel==2.17.0
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
bm25s==0.2.13
boto3==1.38.42
botocore==1.38.42
cachetools==5.5.2
camelot-py==1.0.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromedriver-autoinstaller==0.6.4
click==8.2.1
cloudpathlib==0.21.1
cohere==5.15.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cryptography==45.0.4
cssselect==1.3.0
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
deeplake==3.9.27
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
duckdb==1.3.1
Elixir==0.7.1
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.11.1
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.25.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.2
humbug==0.3.2
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kagglehub==0.3.12
langchain==0.3.26
langchain-community==0.0.38
langchain-core==0.3.68
langchain-google-genai==2.1.6
langchain-openai==0.1.25
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.5
language_data==1.3.0
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.48
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.52.post1
llama-index-embeddings-adapter==0.3.0
llama-index-embeddings-gemini==0.3.2
llama-index-embeddings-openai==0.3.1
llama-index-experimental==0.5.6
llama-index-finetuning==0.3.2
llama-index-graph-stores-nebula==0.5.0
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.2.0
llama-index-legacy==0.9.48.post4
llama-index-llms-azure-openai==0.3.4
llama-index-llms-gemini==0.5.0
llama-index-llms-google-genai==0.2.4
llama-index-llms-mistralai==0.4.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-readers-web==0.4.2
llama-index-readers-wikipedia==0.3.0
llama-index-retrievers-bm25==0.5.2
llama-index-vector-stores-deeplake==0.3.3
llama-index-workflows==1.0.1
llama-parse==0.6.43
llamaindex-py-client==0.1.19
lxml==5.4.0
lxml_html_clean==0.4.2
lz4==4.4.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.12.0
mdurl==0.1.2
mistralai==1.8.2
mistune==3.1.3
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nebula3-python==3.8.3
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
openai==1.93.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
openrouteservice==2.3.3
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
oxylabs==2.0.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.4
pbr==6.1.1
pdfminer.six==20250506
peft==0.15.2
pillow==10.4.0
platformdirs==4.3.8
playwright==1.53.0
pox==0.3.6
ppft==1.7.7
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdf==5.7.0
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
pysbd==0.3.4
PyScreeze==1.0.1
PySocks==1.7.1
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytweening==1.2.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
ragas==0.1.21
RapidFuzz==3.13.0
redlines==0.5.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-click==1.8.9
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.34.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spider-client==0.0.27
SQLAlchemy==2.0.41
sqlalchemy-migrate==0.11.0
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.2
staticvectors==0.2.0
streamlit==1.46.0
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
Tempita==0.6.0
tenacity==8.5.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
trl==0.18.1
typer==0.16.0
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.37.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
xlrd==0.7.1
xlsxwriter==3.2.5
xlwt==0.7.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
